- Fix incorrect comparison of float and double values in `safeAcos` method. - Simplify switches in rounding methods to use `switch` expression. # Fix unclosed condition in `Lwjgl3OpenAlApi` device initialization. + Add `final` keyword to constants in `HelloRawInput`, `JavaAudioSystemLoader`, and `Matrix4f`. - Remove redundant suppress warning annotation in `AssetManager`. # Remove unnecessary float cast in `angle` method of `Vector3f`.

+++ engine/core/src/main/java/mylie/math/Vector3f.java
@@ -60,7 +60,7 @@ public record Vector3f(float x, float y, float z) {
 	}
 
 	public float angle(Vector3f vector) {
-		return (float) Math.acos(dot(vector) / (length() * vector.length()));
+		return Math.acos(dot(vector) / (length() * vector.length()));
 	}
 
 	public Vector3f negate() {
